framework updates

+++ src/TUtil.cc
@@ -959,7 +959,7 @@ TGraph * TUtil::makeCW(double freq,  double amp, double t_min, double t_max, dou
 
   int n=(t_max-t_min)*GSs;
   TGraph * oG=new TGraph(n);
-  int dt=1./GSs;
+  double dt=1./GSs;
   for(int i=0;i<n;i++){
     double t = (double)i*dt;
     double temp=amp*sin(2.*pi*freq*t + phase);
